Fix file names.
