+++ examples/gin/middleware.go
@@ -1,7 +1,6 @@
-package route_stat
+package main
 
 import (
-	"flag"
 	"sync"
 	"time"
 
@@ -9,24 +8,10 @@ import (
 	"github.com/gin-gonic/gin"
 )
 
+var pathMapOnce sync.Once
 var pathMap map[string]string
-var once sync.Once
 
-var env = flag.String("env", "development", "environment, e.g. development or production")
-var host = flag.String("host", "", "host")
-var projectId = flag.Int64("project_id", 0, "project ID")
-var projectKey = flag.String("project_key", "", "project key")
-
-func NewAirbrakeMiddleware(engine *gin.Engine) func(c *gin.Context) {
-	flag.Parse()
-
-	notifier := gobrake.NewNotifierWithOptions(&gobrake.NotifierOptions{
-		ProjectId:   *projectId,
-		ProjectKey:  *projectKey,
-		Host:        *host,
-		Environment: *env,
-	})
-
+func NewAirbrakeMiddleware(engine *gin.Engine, notifier *gobrake.Notifier) func(c *gin.Context) {
 	return func(c *gin.Context) {
 		start := time.Now()
 
@@ -46,7 +31,7 @@ func getRouteName(c *gin.Context, engine *gin.Engine) string {
 }
 
 func extractRouteNames(engine *gin.Engine) {
-	once.Do(func() {
+	pathMapOnce.Do(func() {
 		pathMap = make(map[string]string)
 		for _, ri := range engine.Routes() {
 			pathMap[ri.Handler] = ri.Path
